+++ Sources/CrabArchive/CrabArchive.swift
@@ -1,3 +1,3 @@
 public enum CrabArchive {
-    public static let version = "0.2.7"
+    public static let version = "0.2.8"
 }